e2e_tests/requirements.txt
Appium-Python-Client==3.1.0
selenium==4.15.2
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0
allure-pytest==2.13.2
python-dotenv==1.0.0

